Removing many deprecated methods, moving the sdk2 packages back into the main sdk namespace.
