+++ fh-android-sdk-test/src/main/java/com/feedhenry/sdk/tests/api2/FHAuthRequestTest.java
@@ -23,11 +23,11 @@ import com.feedhenry.sdk.FHActCallback;
 import com.feedhenry.sdk.FHResponse;
 import com.feedhenry.sdk.api.FHAuthRequest;
 import static com.feedhenry.sdk.api.FHAuthSession.SESSION_TOKEN_KEY;
-import com.feedhenry.sdk.api2.FHAuthSession;
+import com.feedhenry.sdk.api.FHAuthSession;
 import com.feedhenry.sdk.tests.MainActivity;
 import com.feedhenry.sdk.tests.sync.FHTestUtils;
 import com.feedhenry.sdk.utils.DataManager;
-import com.feedhenry.sdk2.FHHttpClient;
+import com.feedhenry.sdk.FHHttpClient;
 import cz.msebera.android.httpclient.Header;
 import java.util.concurrent.atomic.AtomicBoolean;
 import org.json.fh.JSONObject;
