+++ fh-android-sdk-test/src/main/java/com/feedhenry/sdk/tests/api2/FHRemoteTest.java
@@ -19,12 +19,11 @@ import android.test.AndroidTestCase;
 import com.feedhenry.sdk.FHActCallback;
 import com.feedhenry.sdk.FHResponse;
 import com.feedhenry.sdk.tests.sync.FHTestUtils;
-import com.feedhenry.sdk2.FHHttpClient;
-import com.feedhenry.sdk2.FHRemote;
+import com.feedhenry.sdk.FHHttpClient;
+import com.feedhenry.sdk.FHRemote;
 import cz.msebera.android.httpclient.Header;
 import org.hamcrest.BaseMatcher;
 import org.hamcrest.Description;
-import org.hamcrest.Matcher;
 import org.json.fh.JSONObject;
 import org.mockito.Matchers;
 import static org.mockito.Matchers.any;
