+++ fh-android-sdk/src/main/java/com/feedhenry/sdk/FHAct.java
@@ -18,19 +18,13 @@ package com.feedhenry.sdk;
 /**
  * Represents a request call to FeedHenry.
  *
- * When asynchronous mode is used, the request will be executed on a separate
- * thread and the calling thread will be returned immediately. When the request
- * is completed, the callback to process the response will be executed on the
- * calling thread using the Looper class. The calling thread needs to be alive
- * for the callback to be processed properly.
+ * The request will be executed on a separate thread and the calling thread will
+ * be returned immediately. When the request is completed, the callback to
+ * process the response will be executed on the calling thread using the Looper
+ * class. The calling thread needs to be alive for the callback to be processed
+ * properly.
  *
- * If you are making API calls on short-lived threads, consider using
- * synchronous mode instead.
- *
- * @deprecated the execute method will become an asynchronous method and the
- * executeAsync methods will be removed in fh-android-sdk:3.0.0/Platform 3.7.0
  */
-@Deprecated
 public interface FHAct {
 
     /**
@@ -46,20 +40,15 @@ public interface FHAct {
      * finishes.
      *
      */
-    void executeAsync();
+    void execute();
 
     /**
      * Executes the request asynchronously. Executes the pCallback function when
      * it finishes.
      *
      * @param pCallback the callback function
-     */
-    void executeAsync(FHActCallback pCallback);
-
-    /**
-     * Executes the request synchronously.
-     *
-     * @param pCallback the callback function
+     * 
      */
     void execute(FHActCallback pCallback);
+
 }
