+++ fh-android-sdk/src/main/java/com/feedhenry/sdk/api/FHAuthSession.java
@@ -26,88 +26,118 @@ import com.feedhenry.sdk.utils.StringUtils;
 import org.json.fh.JSONObject;
 
 public class FHAuthSession {
+
     public static final String SESSION_TOKEN_KEY = "sessionToken";
     private static final String LOG_TAG = "com.feedhenry.sdk.api.FHAuthSession";
 
     private static final String VERIFY_SESSION_ENDPOINT = "admin/authpolicy/verifysession";
     private static final String REVOKE_SESSION_ENDPOINT = "admin/authpolicy/revokesession";
-    private static com.feedhenry.sdk.api2.FHAuthSession instance;
 
-    
-    
-    private FHAuthSession() {
+    private final DataManager mDataManager;
+    private final FHHttpClient mHttpClient;
 
+    /**
+     * Boring constructor with no side effects.
+     *
+     * @param pDataManager a DataManager instance to inject
+     * @param pHttpClient a FHHttpClient to inject
+     */
+    public FHAuthSession(DataManager pDataManager, FHHttpClient pHttpClient) {
+        mDataManager = pDataManager;
+        mHttpClient = pHttpClient;
     }
 
     /**
      * Checks if a sessionToken value exists on the device.
      *
      * @return if the sessionToken exists
-     * @deprecated please use com.feedhenry.sdk.api2.FHAuthSession.exists instead
      */
-    @Deprecated
-    public static boolean exists() {
-        return getInstance(DataManager.getInstance()).exists();
+    public boolean exists() {
+        return mDataManager.read(SESSION_TOKEN_KEY) != null;
     }
 
     /**
      * Gets the value of the current session token.
      *
      * @return the current session token value
-     * @deprecated please use com.feedhenry.sdk.api2.FHAuthSession.getToken instead
      */
-    @Deprecated
-    public static String getToken() {
-        return getInstance(DataManager.getInstance()).getToken();
+    public String getToken() {
+        return mDataManager.read(SESSION_TOKEN_KEY);
     }
 
     /**
      * Saves the seesionToken value on the device.
      *
      * @param sessionToken Session token
-     * @deprecated please use com.feedhenry.sdk.api2.FHAuthSession.save instead
      */
-    @Deprecated
-    protected static void save(String sessionToken) {
-        getInstance(DataManager.getInstance()).save(sessionToken);
+    public void save(String sessionToken) {
+        mDataManager.save(SESSION_TOKEN_KEY, sessionToken);
     }
 
     /**
-     * Calls the remote server to check if the existing sessionToken is actually valid.
+     * Calls the remote server to check if the existing sessionToken is actually
+     * valid.
      *
      * @param pCallback a callback to be executed when remote call is completed
-     * @param pSync     A flag to call it sync
-     * 
-     * @deprecated please use com.feedhenry.sdk.api2.FHAuthSession.verify instead
+     * @param pSync A flag to call it sync
+     *
      */
-    @Deprecated
-    public static void verify(Callback pCallback, boolean pSync) {
-        getInstance(DataManager.getInstance()).verify(pCallback, pSync);
+    public void verify(Callback pCallback, boolean pSync) {
+        String sessionToken = mDataManager.read(SESSION_TOKEN_KEY);
+        if (sessionToken != null) {
+            callRemote(VERIFY_SESSION_ENDPOINT, sessionToken, pCallback, pSync);
+        }
     }
 
     /**
-     * Removes the session token on the device and tries to remove it remotely as well.
+     * Removes the session token on the device and tries to remove it remotely
+     * as well.
      *
-     * @param pSync     A flag to call it sync
-     * 
-     * @deprecated please use com.feedhenry.sdk.api2.FHAuthSession.clear instead 
+     * @param pSync A flag to call it sync
+     *
      */
-    @Deprecated
-    public static void clear(boolean pSync) {
-        getInstance(DataManager.getInstance()).clear(pSync);
+    public void clear(boolean pSync) {
+        String sessionToken = mDataManager.read(SESSION_TOKEN_KEY);
+        if (sessionToken != null) {
+            mDataManager.remove(SESSION_TOKEN_KEY);
+            callRemote(REVOKE_SESSION_ENDPOINT, sessionToken, null, pSync);
+        }
     }
 
-    private static synchronized com.feedhenry.sdk.api2.FHAuthSession getInstance(DataManager dataManager) {
-        if (instance == null) {
-            instance = new com.feedhenry.sdk.api2.FHAuthSession(dataManager, new com.feedhenry.sdk2.FHHttpClient());
-        }
-        return instance;
+    private void callRemote(String pPath, String pSessionToken, final Callback pCallback, boolean pUseSync) {
+        String host = AppProps.getInstance().getHost();
+        String url = StringUtils.removeTrailingSlash(host) + FHRemote.PATH_PREFIX + pPath;
+        JSONObject params = new JSONObject().put(SESSION_TOKEN_KEY, pSessionToken);
+
+        mHttpClient.post(
+                url,
+                null,
+                params,
+                new FHActCallback() {
+            @Override
+            public void success(FHResponse pResponse) {
+                JSONObject res = pResponse.getJson();
+                if (pCallback != null) {
+                    pCallback.handleSuccess(res.getBoolean("isValid"));
+                }
+            }
+
+            @Override
+            public void fail(FHResponse pResponse) {
+                FHLog.w(LOG_TAG, pResponse.getRawResponse());
+                if (pCallback != null) {
+                    pCallback.handleError(pResponse);
+                }
+            }
+        },
+                pUseSync);
     }
-    
+
     public interface Callback {
 
         void handleSuccess(boolean isValid);
 
         void handleError(FHResponse pRes);
     }
+
 }
